+++ ZScriptTests/Runtime/ArrayTests.cs
@@ -31,9 +31,9 @@ namespace ZScriptTests.Runtime
             Assert.IsInstanceOfType(memory.GetVariable("a"), typeof(IList), "The list created by the script is not of the expected type");
             var list = (IList)memory.GetVariable("a");
 
-            Assert.AreEqual(0, list[0], "The list did not have the expected values");
-            Assert.AreEqual(1, list[1], "The list did not have the expected values");
-            Assert.AreEqual(2, list[2], "The list did not have the expected values");
+            Assert.AreEqual((long)0, list[0], "The list did not have the expected values");
+            Assert.AreEqual((long)1, list[1], "The list did not have the expected values");
+            Assert.AreEqual((long)2, list[2], "The list did not have the expected values");
         }
 
         /// <summary>
@@ -59,9 +59,9 @@ namespace ZScriptTests.Runtime
             var list = (IList)memory.GetVariable("a");
 
             Assert.IsInstanceOfType(list[0], typeof(IList), "The list created by the script is not of the expected type");
-            Assert.AreEqual(0, ((IList)list[0])[0], "The list did not have the expected values");
-            Assert.AreEqual(1, list[1], "The list did not have the expected values");
-            Assert.AreEqual(2, list[2], "The list did not have the expected values");
+            Assert.AreEqual((long)0, ((IList)list[0])[0], "The list did not have the expected values");
+            Assert.AreEqual((long)1, list[1], "The list did not have the expected values");
+            Assert.AreEqual((long)2, list[2], "The list did not have the expected values");
         }
 
         /// <summary>
@@ -83,7 +83,7 @@ namespace ZScriptTests.Runtime
             runtime.CallFunction("funca");
 
             // Assert the correct call was made
-            Assert.AreEqual(0, memory.GetVariable("a"), "The list did not have the expected values");
+            Assert.AreEqual((long)0, memory.GetVariable("a"), "The list did not have the expected values");
         }
 
         /// <summary>
@@ -105,7 +105,7 @@ namespace ZScriptTests.Runtime
             runtime.CallFunction("funca");
 
             // Assert the correct call was made
-            Assert.AreEqual(1, memory.GetVariable("a"), "The list did not have the expected values");
+            Assert.AreEqual((long)1, memory.GetVariable("a"), "The list did not have the expected values");
         }
 
         /// <summary>
@@ -129,7 +129,7 @@ namespace ZScriptTests.Runtime
             // Assert the correct call was made
             var list = (IList)memory.GetVariable("a");
 
-            Assert.AreEqual(1, list[0], "The list created by the script is not of the expected type");
+            Assert.AreEqual((long)1, list[0], "The list created by the script is not of the expected type");
         }
 
         /// <summary>
@@ -153,7 +153,7 @@ namespace ZScriptTests.Runtime
             // Assert the correct call was made
             var list = (IList)memory.GetVariable("a");
 
-            Assert.AreEqual(2, ((IList)list[0])[1], "The list created by the script is not of the expected type");
+            Assert.AreEqual((long)2, ((IList)list[0])[1], "The list created by the script is not of the expected type");
         }
     }
 }
