+++ ZScriptTests/Runtime/SubscriptingTests.cs
@@ -27,11 +27,11 @@ namespace ZScriptTests.Runtime
             var memory = runtime.GlobalMemory;
 
             // Set the dictionary on memory now
-            memory.SetVariable("b", new Dictionary<string, object> { { "abc", 10 } });
+            memory.SetVariable("b", new Dictionary<string, object> { { "abc", (long)10 } });
 
             runtime.CallFunction("funca");
 
-            Assert.AreEqual(10, memory.GetVariable("a"), "The VM failed to fetch the correct subscription for the dictionary object stored");
+            Assert.AreEqual((long)10, memory.GetVariable("a"), "The VM failed to fetch the correct subscription for the dictionary object stored");
         }
 
         /// <summary>
@@ -57,7 +57,7 @@ namespace ZScriptTests.Runtime
 
             runtime.CallFunction("funca");
 
-            Assert.AreEqual(10, dict["abc"], "The VM failed to set the correct subscription for the dictionary object stored");
+            Assert.AreEqual((long)10, dict["abc"], "The VM failed to set the correct subscription for the dictionary object stored");
         }
 
         /// <summary>
@@ -83,7 +83,7 @@ namespace ZScriptTests.Runtime
 
             runtime.CallFunction("funca");
 
-            Assert.AreEqual(20, dict["abc"], "The VM failed to set the correct subscription for the dictionary object stored");
+            Assert.AreEqual((long)20, dict["abc"], "The VM failed to set the correct subscription for the dictionary object stored");
         }
     }
 }
